+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2026-05-02
+
+### Changed
+
+- All advancements are now unlocked by default when in Edit Mode.
+- Vastly improved advancement reload speeds.
+- Improved advancement deleting.
+
+### Added
+
+- Added option for reward claiming (default: false).
+
 ## [1.2.0] - 2026-05-02
 
 ### Added